refactor: remove unnecessary parent component
